+++ package.json
@@ -37,6 +37,7 @@
   "homepage": "https://github.com/zabo-api/zabo-sdk-js#readme",
   "devDependencies": {
     "@babel/core": "^7.7.2",
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
     "babel-loader": "^8.0.6",
     "mocha": "^6.2.0",
     "should": "^13.2.3",
